refactor auth options and fix build
