+++ lib/auth.ts
@@ -1,0 +1,30 @@
+import { NextAuthOptions } from 'next-auth';
+import Credentials from 'next-auth/providers/credentials';
+import { PrismaAdapter } from '@next-auth/prisma-adapter';
+import { prisma } from '@/lib/db';
+import { compare } from 'bcryptjs';
+import { z } from 'zod';
+
+export const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  session: { strategy: 'jwt' },
+  providers: [
+    Credentials({
+      credentials: {
+        email: { label: 'Email', type: 'email' },
+        password: { label: 'Password', type: 'password' }
+      },
+      async authorize(credentials) {
+        const parsed = z
+          .object({ email: z.string().email(), password: z.string().min(1) })
+          .safeParse(credentials);
+        if (!parsed.success) return null;
+        const user = await prisma.user.findUnique({ where: { email: parsed.data.email } });
+        if (!user || !user.passwordHash) return null;
+        const valid = await compare(parsed.data.password, user.passwordHash);
+        if (!valid) return null;
+        return { id: user.id, email: user.email, role: user.role };
+      }
+    })
+  ]
+};
