+++ package.json
@@ -52,6 +52,7 @@
     "@testing-library/jest-dom": "6.4.2",
     "@testing-library/react": "14.1.2",
     "@testing-library/user-event": "14.4.3",
+    "@types/nodemailer": "^6.4.8",
     "eslint": "8.57.0",
     "eslint-config-next": "14.2.5",
     "eslint-config-prettier": "9.1.0",
